Adds a workaround for https://github.com/benchflow/client/issues/7
There is currently an issue when passing to Docker a volume that is
shared with the host operating system by VirtualBox. You end up sharing
previous version of files as discussed on the following and linked
issue: https://github.com/benchflow/client/issues/7. This is a
workaround for the discussed issue, waiting for an upstream update from
the VirtualBox community on this topic.